Fixed sizet warnings (#101)
* don't cycle playlist if on "street fighter and capcom fighters" or "street fighter"

* lower errors to warnings in logs

* - fixed sort to be more supportive of sort types (monkofthefunk)
 - fixed playcount as an reloadableimage (monkofthfunk)
 - improve restart video fast scroll flicker (monkofthefunk)

* fixed reload playcount after launch for images

* linuxfix

* fix

* fixed high GPU due to marquee rendering thread on game launch

* revert changes

* fix size_t warnings

* fix exception

+++ RetroFE/Source/Graphics/Animate/Animation.cpp
@@ -62,7 +62,7 @@ TweenSet *Animation::tweenSet(unsigned int index)
 }
 
 
-unsigned int Animation::size()
+size_t Animation::size()
 {
     return animationVector_.size();
 }
